a2a/requirements.txt
a2a-sdk
httpx
uvicorn
python-dotenv